cleaned up View Comment button code

+++ src/components/Post.js
@@ -10,9 +10,13 @@ const Post = (props) => {
     
     const handleDelete = () => { //on form submit state sent to addPost action
        
-        props.deletePost(props.match.params.id)
-        props.history.push('/posts')
-        props.history.go(0)
+        if (window.confirm('Are you sure you want to delete this post?')) {
+            props.deletePost(props.match.params.id)
+            props.history.push('/posts')
+            props.history.go(0)
+        } else {
+            console.log('post was not deleted');
+        }
     }
 
     const [drop, setDrop] = useState(false);
@@ -24,7 +28,6 @@ const Post = (props) => {
     //     console.log(drop);
     // }, [drop]);
 
-    if (!drop){
     return (
         <div className='post'>
             <h2>{post ? post.title : null}</h2>
@@ -38,33 +41,12 @@ const Post = (props) => {
             <button className='button' type='delete' onClick={handleDelete}>Delete</button>
             <br></br>
             <button className='comment-button' onClick={buttonHandler} type="button">
-                <i className="far fa-comment-dots">View Comments</i>            
+                <i className="far fa-comment-dots">{!drop ? 'View Comments' : 'Hide Comments'}</i>            
             </button>
+            {drop && (<CommentsContainer post={post}/>)}
         </div>
         //if Post exists then render, otherwise null
     );
-    } else {
-        return (
-            <div className='post'>
-                <h2>{post ? post.title : null}</h2>
-                <img className='post-image' src={post ? post.image : null} alt=''/>
-                <hr></hr>
-                <br></br>
-                <p>{post ? post.body : null}</p>
-                <br></br>
-                <hr></hr>
-                <br></br>
-                <button className='button' type='delete' onClick={handleDelete}>Delete</button>
-                <br></br>
-                <button className='comment-button' onClick={buttonHandler} type="button">
-                    <i className="far fa-comment-dots">Hide Comments</i>            
-                </button>
-                <CommentsContainer post={post}/>
-            </div>
-            //if Post exists then render, otherwise null
-        );
-    }
-    
 }
 
 export default connect(null, {deletePost})(Post)
